[REFACTOR(uimodel)] uimodel 패키지 생성 및 model 이동
