+++ app/src/main/java/woowacourse/movie/ui/seat/uimodel/SelectedSeatsModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.ui.seat.uimodel
 
 import java.io.Serializable
